doc: change folder structure
