+++ template/apps/api/src/config/environment/development.json
@@ -5,27 +5,5 @@
   },
   "apiUrl": "http://localhost:3001",
   "webUrl": "http://localhost:3002",
-  "redis": "redis://:@localhost:6379",
-  "docs": {
-    "info": {
-      "title": "My API",
-      "description": "This is the my API for public",
-      "version": "1.0.0",
-      "license": {
-        "name": "MIT",
-        "url": "https://opensource.org/license/mit/"
-      },
-      "contact": {
-        "name": "API Support",
-        "url": "https://www.example.com/support",
-        "email": "support@example.com"
-      }
-    },
-    "servers": [
-      {
-        "url": "https://development.gigantic-server.com/v1",
-        "description": "Development server"
-      }
-    ]
-  }
+  "redis": "redis://:@localhost:6379"
 }
