+++ template/apps/api/src/resources/account/actions/get/index.ts
@@ -1,22 +1,15 @@
 import { AppKoaContext, AppRouter } from 'types';
 
 import { userService } from 'resources/user';
-import { docsUtil } from 'utils';
+import { docsService } from 'services';
+import docConfig from './doc';
 
 async function handler(ctx: AppKoaContext) {
   ctx.body = { ...userService.getPublic(ctx.state.user), isShadow: Boolean(ctx.state.isShadow) };
 }
 
 export default (router: AppRouter) => {
-  docsUtil.registerDocs({
-    private: true,
-    tags: ['account'],
-    method: 'get',
-    path: '/account/',
-    summary: 'Get current user',
-    request: {},
-    responses: {},
-  });
+  docsService.registerDocs(docConfig);
 
   router.get('/', handler);
 };
