+++ template/apps/api/src/resources/account/actions/resend-email/index.ts
@@ -1,15 +1,13 @@
 import { z } from 'zod';
 
 import config from 'config';
-import { docsUtil, securityUtil } from 'utils';
-import { emailService } from 'services';
+import { securityUtil } from 'utils';
+import { emailService, docsService } from 'services';
 import { validateMiddleware } from 'middlewares';
 import { AppKoaContext, Next, AppRouter } from 'types';
 import { userService, User } from 'resources/user';
-
-const schema = z.object({
-  email: z.string().min(1, 'Please enter email').email('Email format is incorrect.'),
-});
+import docConfig from './doc';
+import { schema } from './schema';
 
 interface ValidatedData extends z.infer<typeof schema> {
   user: User;
@@ -42,17 +40,7 @@ async function handler(ctx: AppKoaContext<ValidatedData>) {
 }
 
 export default (router: AppRouter) => {
-  docsUtil.registerDocs({
-    private: false,
-    tags: ['account'],
-    method: 'post',
-    path: '/account/resend-email',
-    summary: 'Resend email',
-    request: {
-      body: { content: { 'application/json': { schema } } },
-    },
-    responses: {},
-  });
+  docsService.registerDocs(docConfig);
 
   router.post('/resend-email', validateMiddleware(schema), validator, handler);
 };
