+++ template/apps/api/src/resources/account/actions/resend-email/schema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+  email: z.string().min(1, 'Please enter email').email('Email format is incorrect.'),
+});
