+++ template/apps/api/src/resources/account/actions/reset-password/index.ts
@@ -1,17 +1,12 @@
 import { z } from 'zod';
 
-import { docsUtil, securityUtil } from 'utils';
+import { securityUtil } from 'utils';
+import { docsService } from 'services';
 import { validateMiddleware } from 'middlewares';
 import { AppKoaContext, Next, AppRouter } from 'types';
 import { userService, User } from 'resources/user';
-
-const schema = z.object({
-  token: z.string().min(1, 'Token is required'),
-  password: z.string().regex(
-    /^(?=.*[a-z])(?=.*\d)[A-Za-z\d\W]{6,}$/g,
-    'The password must contain 6 or more characters with at least one letter (a-z) and one number (0-9).',
-  ),
-});
+import docConfig from './doc';
+import { schema } from './schema';
 
 interface ValidatedData extends z.infer<typeof schema> {
   user: User;
@@ -42,18 +37,7 @@ async function handler(ctx: AppKoaContext<ValidatedData>) {
 }
 
 export default (router: AppRouter) => {
-  docsUtil.registerDocs({
-    private: false,
-    tags: ['account'],
-    method: 'put',
-    path: '/account/reset-password',
-    summary: 'Reset password',
-    description: 'Just reset users password',
-    request: {
-      body: { content: { 'application/json': { schema } } },
-    },
-    responses: {},
-  });
+  docsService.registerDocs(docConfig);
 
   router.put('/reset-password', validateMiddleware(schema), validator, handler);
 };
