+++ template/apps/api/src/resources/account/actions/shadow-login/schema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+  id: z.string().min(1, 'Id is required'),
+});
