+++ template/apps/api/src/resources/account/actions/sign-in/schema.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+  email: z.string().min(1, 'Please enter email').email('Email format is incorrect.'),
+  password: z.string().min(1, 'Please enter password'),
+});
