+++ template/apps/api/src/resources/account/actions/sign-out/index.ts
@@ -1,0 +1,15 @@
+import { authService, docsService } from 'services';
+import { AppKoaContext, AppRouter } from 'types';
+import docConfig from './doc';
+
+const handler = async (ctx: AppKoaContext) => {
+  await authService.unsetTokens(ctx);
+
+  ctx.body = {};
+};
+
+export default (router: AppRouter) => {
+  docsService.registerDocs(docConfig);
+
+  router.post('/sign-out', handler);
+};
