+++ template/apps/api/src/resources/account/actions/update/doc.ts
@@ -1,0 +1,17 @@
+import { resourceName } from '../../constants';
+import { schema } from './schema';
+import { RouteExtendedConfig } from 'services/docs.service';
+
+const config: RouteExtendedConfig = {
+  private: true,
+  tags: [resourceName],
+  method: 'put',
+  path: `/${resourceName}/`,
+  summary: 'Update user profile',
+  request: {
+    body: { content: { 'application/json': { schema } } },
+  },
+  responses: {},
+};
+
+export default config;
