+++ template/apps/api/src/resources/account/actions/update/schema.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+  firstName: z.string().min(1, 'Please enter First name').max(100),
+  lastName: z.string().min(1, 'Please enter Last name').max(100),
+  password: z.string().regex(
+    /^$|^(?=.*[a-z])(?=.*\d)[A-Za-z\d\W]{6,}$/g,
+    'The password must contain 6 or more characters with at least one letter (a-z) and one number (0-9).',
+  ),
+});
