+++ template/apps/api/src/resources/account/actions/verify-email/index.ts
@@ -6,11 +6,9 @@ import { AppKoaContext, Next, AppRouter } from 'types';
 import { validateMiddleware } from 'middlewares';
 import { authService, emailService } from 'services';
 import { userService, User } from 'resources/user';
-import { docsUtil } from 'utils';
-
-const schema = z.object({
-  token: z.string().min(1, 'Token is required'),
-});
+import { docsService } from 'services';
+import { schema } from './schema';
+import docConfig from './doc';
 
 interface ValidatedData extends z.infer<typeof schema> {
   user: User;
@@ -51,21 +49,7 @@ async function handler(ctx: AppKoaContext<ValidatedData>) {
 }
 
 export default (router: AppRouter) => {
-  docsUtil.registerDocs({
-    private: false,
-    tags: ['account'],
-    method: 'get',
-    path: '/account/verify-email',
-    summary: 'Verify email',
-    request: {
-      query: schema,
-    },
-    responses: {
-      302: {
-        description: 'Redirect to web app',
-      },
-    },
-  });
+  docsService.registerDocs(docConfig);
 
   router.get('/verify-email', validateMiddleware(schema), validator, handler);
 };
