+++ template/apps/api/src/resources/account/actions/verify-email/schema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+  token: z.string().min(1, 'Token is required'),
+});
