+++ template/apps/api/src/resources/account/actions/verify-reset-token/index.ts
@@ -4,12 +4,9 @@ import config from 'config';
 import { validateMiddleware } from 'middlewares';
 import { AppKoaContext, AppRouter } from 'types';
 import { User, userService } from 'resources/user';
-import { docsUtil } from 'utils';
-
-const schema = z.object({
-  email: z.string().min(1, 'Please enter email').email('Email format is incorrect.'),
-  token: z.string().min(1, 'Token is required'),
-});
+import { docsService } from 'services';
+import { schema } from './schema';
+import docConfig from './doc';
 
 interface ValidatedData extends z.infer<typeof schema> {
   user: User;
@@ -28,22 +25,7 @@ async function validator(ctx: AppKoaContext<ValidatedData>) {
 }
 
 export default (router: AppRouter) => {
-  docsUtil.registerDocs({
-    private: false,
-    tags: ['account'],
-    method: 'get',
-    path: '/account/verify-reset-token',
-    summary: 'Verify reset token',
-    description: 'Check reset token',
-    request: {
-      query: schema,
-    },
-    responses: {
-      302: {
-        description: 'Redirect to reset password page',
-      },
-    },
-  });
+  docsService.registerDocs(docConfig);
 
   router.get('/verify-reset-token', validateMiddleware(schema), validator);
 };
