+++ template/apps/api/src/resources/account/constants.ts
@@ -1,0 +1,1 @@
+export const resourceName = 'account';
