+++ template/apps/api/src/resources/user/actions/list/doc.ts
@@ -1,0 +1,17 @@
+import { resourceName } from '../../constants';
+import { schema } from './schema';
+import { RouteExtendedConfig } from 'services/docs.service';
+
+const config: RouteExtendedConfig = {
+  private: true,
+  tags: [resourceName],
+  method: 'get',
+  path: `/${resourceName}/`,
+  summary: 'List of users',
+  request: {
+    query: schema,
+  },
+  responses: {},
+};
+
+export default config;
