+++ template/apps/api/src/resources/user/actions/list/index.ts
@@ -4,22 +4,9 @@ import { SortDirection } from '@paralect/node-mongo';
 import { AppKoaContext, AppRouter } from 'types';
 import { validateMiddleware } from 'middlewares';
 import { userService } from 'resources/user';
-import { docsUtil } from 'utils';
-
-const schema = z.object({
-  page: z.string().transform(Number).default('1').openapi({ type: 'string' }),
-  perPage: z.string().transform(Number).default('10').openapi({ type: 'string' }),
-  sort: z.object({
-    createdOn: z.string(),
-  }).default({ createdOn: 'desc' }),
-  filter: z.object({
-    createdOn: z.object({
-      sinceDate: z.string(),
-      dueDate: z.string(),
-    }).nullable().default(null),
-  }).nullable().default(null),
-  searchValue: z.string().default(''),
-});
+import { docsService } from 'services';
+import { schema } from './schema';
+import docConfig from './doc';
 
 type ValidatedData = Omit<z.infer<typeof schema>, 'sort'> & {
   sort: { [name: string]: SortDirection };
@@ -64,18 +51,7 @@ async function handler(ctx: AppKoaContext<ValidatedData>) {
 }
 
 export default (router: AppRouter) => {
-  docsUtil.registerDocs({
-    private: true,
-    tags: ['users'],
-    method: 'get',
-    path: '/users/',
-    summary: 'List of users',
-    description: 'Get all users in database paginated',
-    request: {
-      query: schema,
-    },
-    responses: {},
-  });
+  docsService.registerDocs(docConfig);
 
   router.get('/', validateMiddleware(schema), handler);
 };
