+++ template/apps/api/src/resources/user/actions/update/index.ts
@@ -3,12 +3,7 @@ import { z } from 'zod';
 import { AppKoaContext, Next, AppRouter } from 'types';
 import { validateMiddleware } from 'middlewares';
 import { userService } from 'resources/user';
-
-const schema = z.object({
-  firstName: z.string().min(1, 'Please enter First name').max(100),
-  lastName: z.string().min(1, 'Please enter Last name').max(100),
-  email: z.string().min(1, 'Please enter email').email('Email format is incorrect.'),
-});
+import { schema } from './schema';
 
 type ValidatedData = z.infer<typeof schema>;
 type Request = {
