+++ template/apps/api/src/resources/user/actions/update/schema.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+  firstName: z.string().min(1, 'Please enter First name').max(100),
+  lastName: z.string().min(1, 'Please enter Last name').max(100),
+  email: z.string().min(1, 'Please enter email').email('Email format is incorrect.'),
+});
