+++ template/apps/api/src/resources/user/constants.ts
@@ -1,0 +1,1 @@
+export const resourceName = 'users';
