+++ template/apps/api/src/services/docs.service.ts
@@ -1,7 +1,25 @@
 import { z } from 'zod';
-import { omit } from 'lodash';
+import { omit, capitalize } from 'lodash';
 import { default as projectConfig } from 'config';
 import { OpenAPIGenerator, OpenAPIRegistry, RouteConfig, extendZodWithOpenApi } from '@asteasolutions/zod-to-openapi';
+
+const DEFAULT_CONFIG = {
+  info: {
+    title: 'My API',
+    description: 'This is the my API for public',
+    version: '1.0.0',
+    license: {
+      name: 'MIT',
+      url: 'https://opensource.org/license/mit/',
+    },
+  },
+  servers: [
+    {
+      url: projectConfig.apiUrl,
+      description: `${capitalize(projectConfig.env)} server`,
+    },
+  ],
+};
 
 let registry: OpenAPIRegistry;
 const initClient = () => {
@@ -9,14 +27,10 @@ const initClient = () => {
   registry = new OpenAPIRegistry();
 };
 
-interface RouteExtendedConfig extends Omit<RouteConfig, 'security'> {
+export interface RouteExtendedConfig extends Omit<RouteConfig, 'security'> {
   private: boolean,
 }
 const registerDocs = (config: RouteExtendedConfig): void => {
-  if (!projectConfig.docs) {
-    return;
-  }
-
   if (!registry) {
     throw Error('OpenAPIRegistry is not initialized');
   }
@@ -37,17 +51,13 @@ const registerDocs = (config: RouteExtendedConfig): void => {
 };
 
 const getDocs = () => {
-  if (!projectConfig.docs) {
-    return {};
-  }
-
   if (!registry) {
     throw Error('OpenAPIRegistry is not initialized');
   }
 
   const generator = new OpenAPIGenerator(registry.definitions, '3.1.0');
 
-  return generator.generateDocument(projectConfig.docs as any);
+  return generator.generateDocument(DEFAULT_CONFIG);
 };
 
 export default {
