+++ template/apps/api/src/utils/index.ts
@@ -2,12 +2,10 @@ import configUtil from './config.util';
 import securityUtil from './security.util';
 import routeUtil from './routes.util';
 import promiseUtil from './promise.util';
-import docsUtil from './docs.util';
 
 export {
   configUtil,
   securityUtil,
   routeUtil,
   promiseUtil,
-  docsUtil,
 };
